% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets_geofabrik.R
\docType{data}
\name{BE_OSM_LANDUSE}
\alias{BE_OSM_LANDUSE}
\title{BE_OSM_LANDUSE}
\description{
Geofabrik shapefile landuse imported as: BE_OSM_LANDUSE. More information about this data can be found in the inst/docs folder
}
\examples{
\dontrun{
data(BE_OSM_LANDUSE)
str(as.data.frame(BE_OSM_LANDUSE))
library(sp)
plot(BE_OSM_LANDUSE)
}
}
\references{
\url{https://taginfo.openstreetmap.org/}, \url{http://wiki.openstreetmap.org/wiki/Tags}, \url{http://download.geofabrik.de/europe/belgium-latest.shp.zip}, \url{http://download.geofabrik.de/europe/belgium.html}
}

