% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{BelgiumMapsOpenStreetMap-package}
\alias{BelgiumMapsOpenStreetMap}
\alias{BelgiumMapsOpenStreetMap-package}
\title{Maps of Belgium extracted from OpenStreetMap}
\description{
Maps of Belgium extracted from OpenStreetMap
}
\examples{
## Administrative areas
data(BE_OSM_ADMIN)
str(as.data.frame(BE_OSM_ADMIN))

library(sp)
plot(BE_OSM_ADMIN)
plot(subset(BE_OSM_ADMIN, TAG.admin.level \%in\% "2"))
plot(subset(BE_OSM_ADMIN, TAG.admin.level \%in\% "4"))
plot(subset(BE_OSM_ADMIN, TAG.admin.level \%in\% "6"))
plot(subset(BE_OSM_ADMIN, TAG.admin.level \%in\% "7"))
plot(subset(BE_OSM_ADMIN, TAG.admin.level \%in\% "8"))
plot(subset(BE_OSM_ADMIN, TAG.admin.level \%in\% "9"))
}

