% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.R
\name{tag_extractor}
\alias{tag_extractor}
\title{Extract tag elements from the TAGS field of BE_OSM_ADMIN}
\usage{
tag_extractor(x, tag)
}
\arguments{
\item{x}{a character vector}

\item{tag}{a character string with the tag to extract}
}
\value{
a character vector with the extracted tags values
}
\description{
Extract tag elements from the TAGS field of BE_OSM_ADMIN
}
\examples{
data(BE_OSM_ADMIN)
mytags <-  sapply(BE_OSM_ADMIN$TAGS, FUN=tag_extractor, tag = "ref:INS")
}

